Refactor diagnoses Enums to sets (no testcases)
Diagnoses use Python sets instead of Enums now.
Program no longer halts when an undefined diagnosis encountered.
Instead, it is assumed to be valid and query runs as normal.

A JSON file can be provided for diagnoses aliases, which allow
different inputs to be recognized as the same diagnosis.
This property holds for diagnoses filters and CSV ticket info.

e.g. "hdmiother" would still file under "Other HDMI"
if the specified JSON includes this definition.

Does not include test cases. To add in next commit.