Keep originals in ticket.diagnoses (no testcases)
Fix issue whereby ticket.diagnoses would keep a canonical version of a
diagnosis string if that string did not map to a display name from a
diagnoses aliases file.

Now, strings which do not map to a display name are simply kept exactly
as they appeared in the ticket report.

This also required keeping Ticket.diagnoses as a list, undoing the
previous change whereby it was kept as a set. Now, conversion of
diagnoses lists to sets is only performed in filter_tickets() in
organization.py.

Test cases to be included in a future commit.